+++ services/kms/v1api/wait/wait_test.go
@@ -122,7 +122,7 @@ func fixtureWrappingKey(state string) *kms.WrappingKey {
 		Purpose:     kms.WRAPPINGPURPOSE_WRAP_SYMMETRIC_KEY,
 		State:       state,
 		ExpiresAt:   testDate,
-		PublicKey:   testPublicKey,
+		PublicKey:   utils.Ptr(testPublicKey),
 	}
 }