+++ services/kms/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v1.6.0
+- `v1api`: **Breaking change:** Change type of field `PublicKey` from `string` to `*string` in `WrappingKey` model struct
+- Deprecated SDK layer in root of the module: The field `PublicKey` in the `WrappingKey` model struct is not marked as required anymore
+
 ## v1.5.1
 - **Dependencies:** Bump STACKIT SDK core module from `v0.22.0` to `v0.23.0`